+++ src/modules/page/GangHitlistHelper.js
@@ -1,5 +1,6 @@
 const GangHitlistHelper = {
     cmds: ['gang', 'gang2'],
+    staff: false,
     settings: [
         { key: 'GangHitlistHelper_HitlistPageTracker', label: 'Hitlist Page Tracker' },
         { key: 'GangHitlistHelper_HitlistMarkRed', label: 'Hitlist Mark Red' },