+++ src/modules/page/HitlistHelper.js
@@ -1,5 +1,6 @@
 const HitlistHelper = {
     cmds: 'battle',
+    staff: false,
     settings: [
         { key: 'HitlistHelper_HighlightOnline', label: 'Highlight Online Players' },
         { key: 'HitlistHelper_RememberSort', label: 'Enable Client-side Sorting & Remember' }