+++ src/modules/page/KurtzCampHelper.js
@@ -1,5 +1,6 @@
 const KurtzCampHelper = {
     cmds: 'camp_kurtz',
+    staff: false,
     init: function() {
         // Check Settings
         const settings = Utils.getSettings();