+++ src/modules/page/LivingAreaHelper.js
@@ -1,5 +1,6 @@
 const LivingAreaHelper = {
     cmds: '',
+    staff: false,
     settings: [
         { key: 'LivingAreaHelper_StatRatioTracker', label: 'Stat Ratio Tracker' },
         { key: 'LivingAreaHelper_CopyStatsBtn', label: 'Copy Stats Button' },